requirements.txt
certifi==2019.11.28
cffi==1.14.0
cryptography==2.8
decorator==4.4.2
future==0.18.2
gitdb==4.0.2
gitdb2==4.0.2
GitPython==3.0.8
pycparser==2.20
PySocks==1.7.1
python-telegram-bot==12.4.2
six==1.14.0
smmap==3.0.1
tornado==6.0.4